+++ resources/lang/en/admin/planet.php
@@ -1,0 +1,13 @@
+<?php
+
+return [
+    'index' => [
+        'id' => 'ID',
+        'galaxy' => 'Galaxy',
+        'name' => 'Name',
+        'planet' => 'Planet',
+        'system' => 'System',
+        'title' => 'Planets list',
+        'total' => 'Total',
+    ],
+];