+++ resources/views/admin/partials/menu_super_operator.blade.php
@@ -11,7 +11,7 @@
     <tr><td><a href="paneladmina.php" accesskey="k">@lang('admin/menu.players-search')</a></td></tr>
     <tr><td><a href="{{ route('admin.resource.add') }}" accesskey="k">@lang('admin/menu.resources-add')</a></td></tr>
     <tr><td><a href="{{ route('admin.resource.subtract') }}" accesskey="k">@lang('admin/menu.resources-sub')</a></td></tr>
-    <tr><td class="separated"><a href="colonies.php" accesskey="1">@lang('admin/menu.planets')</a></td></tr>
+    <tr><td class="separated"><a href="{{ route('admin.planets') }}" accesskey="1">@lang('admin/menu.planets')</a></td></tr>
     <tr><td><a href="activeplanet.php" accesskey="k">@lang('admin/menu.planets-active')</a></td></tr>
     <tr><td><a href="{{ route('admin.moons') }}" accesskey="k">@lang('admin/menu.moons')</a></td></tr>
     <tr><td><a href="add_moon.php" accesskey="k">@lang('admin/menu.moon-add')</a></td></tr>